+++ components/form/admin/types/rating.type.tsx
@@ -14,8 +14,10 @@ export const RatingType: React.FC<AdminFieldTypeProps> = (props) => {
         name={[props.field.name as string, 'value']}
         labelCol={{ span: 6 }}
         extra={t('type:rating.clearNote')}
-        getValueFromEvent={(value: number) => (value !== undefined ? value.toFixed(1) : undefined)}
-        getValueProps={(value: string) => ({ value: parseFloat(value) })}
+        getValueFromEvent={(value: number) =>
+          typeof value === 'number' ? value.toFixed(2) : value
+        }
+        getValueProps={(value: string) => ({ value: value ? parseFloat(value) : undefined })}
       >
         <Rate allowHalf allowClear />
       </Form.Item>